Add last_update table to screener

+++ backend/api/technical_screener.py
@@ -20,12 +20,12 @@ class RESTTechnicaScreener(Resource):
     def get(self):
         pattern = request.args.get("pattern", None)
 
+        if pattern == None:
+            return {"msg": "no pattern"}
+
         existing_tables = {}
         for table in postgres_db.engine.table_names(schema="sp500_companies"):
             existing_tables[table] = True
-
-        if pattern == None:
-            return {"msg": "no pattern"}
 
         # 1. get s&p 500 companies
         stats = BeautifulSoup(
@@ -36,10 +36,50 @@ class RESTTechnicaScreener(Resource):
         sp500_tickers = list(df["Symbol"])
 
         for ticker in sp500_tickers:
-            # 2. lookup data from yfinance once per week
-            lookup_data = False
+            # 2. Update data once per week
+            postgres_db.engine.execute(
+                """
+                    CREATE TABLE IF NOT EXISTS sp500_companies."last_update" (
+                        date DATE NOT NULL
+                    )
+                """
+            )
+
+            last_update = pd.read_sql_table(
+                "last_update", postgres_db.engine, schema="sp500_companies"
+            )
+
+            today = datetime.today().strftime("%Y-%m-%d")
+            lookup_data = False  # Whether or not to look up data from yfinance
+
+            if last_update.empty:
+                postgres_db.engine.execute(
+                    """
+                        INSERT INTO sp500_companies."last_update"
+                        VALUES ('{date}');
+                    """.format(
+                        date=today
+                    )
+                )
+                lookup_data = True
+            else:
+                last_update_value = last_update.iloc[0]["date"].strftime("%Y-%m-%d")
+                one_week_ago = (datetime.today() - relativedelta(days=7)).strftime(
+                    "%Y-%m-%d"
+                )
+                if last_update_value == one_week_ago:
+                    postgres_db.engine.execute(
+                        """
+                        UPDATE sp500_companies."last_update"
+                        SET date = {new_date}
+                        WHERE date={date};
+                    """.format(
+                            new_date=today, date=last_update_value
+                        )
+                    )
+                    lookup_data = True
+
             if lookup_data:
-                today = datetime.today().strftime("%Y-%m-%d")
                 start_date = (datetime.now() - relativedelta(years=1)).strftime(
                     "%Y-%m-%d"
                 )  # start 1 year ago
@@ -58,17 +98,24 @@ class RESTTechnicaScreener(Resource):
                 )
                 df.index.names = ["date"]
 
-                postgres_db.engine.execute(
-                    """
-                        DROP TABLE sp500_companies."{ticker}"
-                    """.format(
-                        ticker=ticker
+                if ticker in existing_tables:
+                    postgres_db.engine.execute(
+                        """
+                            DROP TABLE sp500_companies."{ticker}"
+                        """.format(
+                            ticker=ticker
+                        )
                     )
-                )
+
                 df.to_sql(ticker, postgres_db.engine, schema="sp500_companies")
 
             # 4. read from database as dataframe
+            ticker_data = pd.read_sql_table(
+                ticker, postgres_db.engine, schema="sp500_companies"
+            )
+            print(ticker_data)
 
+            return "Looked up"
             # 5. Look for pattern with :
             #   - pattern_function = getattr(talib, pattern)
             #   - result = pattern_function(df["Open"], df["High"], df["Low"], df["Close"])
